requirements.txt
anaconda-client==1.1.0
beautifulsoup4==4.4.1
bokeh==0.10.0
boto==2.39.0
bz2file==0.98
certifi==2015.9.6.2
cffi==1.3.0
characteristic==14.3.0
clyent==0.4.0
cryptography==1.1
cssselect==0.9.1
cycler==0.9.0
decorator==4.0.6
Flask==0.10.1
gensim==0.12.4
gevent==1.0.2
greenlet==0.4.9
h5py==2.5.0
httpretty==0.8.10
idna==2.0
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.0
itsdangerous==0.24
Jinja2==2.8
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
-e git+https://github.com/fchollet/keras.git@399c00c7f87660f5a660df72fdf6fd1192c32be2#egg=Keras
lxml==3.4.4
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.1
nbconvert==4.0.0
nbformat==4.0.1
networkx==1.10
nose==1.3.7
notebook==4.0.6
numpy==1.10.2
pafy==0.4.3
pandas==0.17.0
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
Pillow==3.1.0
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
Pygments==2.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.0
queuelib==1.4.2
requests==2.8.1
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.1
Scrapy==1.0.3
service-identity==14.0.0
simplegeneric==0.8.1
six==1.10.0
smart-open==1.3.2
smop==0.0.0
terminado==0.5
-e git://github.com/Theano/Theano.git@8020a38fd54ee2aad3d20dcf9c0f1e5e4eac4e76#egg=Theano
tornado==4.2.1
tqdm==2.2.4
traitlets==4.0.0
Twisted==15.4.0
w3lib==1.13.0
Werkzeug==0.10.4
wheel==0.26.0
youtube-dl==2016.2.10
zope.interface==4.1.3
